Finished lab i think
